。。。

+++ background/app/controller/user.js
@@ -13,7 +13,7 @@ class UserController extends Controller {
   }
 
   async getMyInfo() {
-    const uid = this.ctx.user.id;
+    const uid = this.ctx.session.user.id;
     uid || this.ctx.helper.createRes(403, 'User not login QAQ');
     const response = await this.userService.getInfo(uid);
     if (response) {
@@ -34,6 +34,92 @@ class UserController extends Controller {
       this.ctx.helper.createRes(404, 'User is not found QAQ');
     }
   }
+  async updateInfo() {
+    const uid = this.ctx.session.user.id;
+    uid || this.ctx.helper.createRes(203, 'User not login QAQ');
+    const {
+      name,
+      signature,
+      avatar
+    } = this.ctx.request.body;
+    const response = await this.userService.updateInfo(uid,name,signature,avatar);
+    this.ctx.helper.successRes('sucess',response);
+  }
+
+  async thirdParty() {
+    const uid = this.ctx.session.user.id;
+    uid || this.ctx.helper.createRes(203, 'User not login QAQ');
+    const {
+      data
+    } = this.ctx.request.body;
+    const {
+      type
+    } = this.ctx.params;
+    switch(type){
+      case 'bili': 
+      const response = await this.userService.updateBili(uid,data);
+      this.ctx.helper.successRes('sucess',response);
+      break;
+      case 'weibo': 
+      const response = await this.userService.updateWeibo(uid,data);
+      this.ctx.helper.successRes('sucess',response);
+      break;
+      case 'fivesong': 
+      const response = await this.userService.updateFivesong(uid,data);
+      this.ctx.helper.successRes('sucess',response);
+      break;
+      case 'tieba': 
+      const response = await this.userService.updatetieba(uid,data);
+      this.ctx.helper.successRes('sucess',response);
+      break;
+    }
+  }
+
+  async getUpdateNum(){
+    const {
+      uid
+    } = this.ctx.params;
+    if (uid === 'own'){
+      const id = this.ctx.session.user.id;
+      id || this.ctx.helper.createRes(203, 'User not login QAQ');
+    } else{
+      const id = uid
+    }
+    const response = await this.userService.getUpdateNum(uid);
+    this.ctx.helper.successRes('sucess',response);
+  }
+
+  async getCreateNum(){
+    const {
+      uid
+    } = this.ctx.params;
+    if (uid === 'own'){
+      const id = this.ctx.session.user.id;
+      id || this.ctx.helper.createRes(203, 'User not login QAQ');
+    } else{
+      const id = uid
+    }
+    const response = await this.userService.getCreateNum(uid);
+    this.ctx.helper.successRes('sucess',response);
+  }
+
+
+  async getFocusNum(){
+    const {
+      uid
+    } = this.ctx.params;
+    if (uid === 'own'){
+      const id = this.ctx.session.user.id;
+      id || this.ctx.helper.createRes(203, 'User not login QAQ');
+    } else{
+      const id = uid
+    }
+    const response = await this.userService.getFocusNum(uid);
+    this.ctx.helper.successRes('sucess',response);
+  }
+
+
+
 
   async getUserCollection() {
     const {
@@ -42,7 +128,7 @@ class UserController extends Controller {
     const {
       offset = DEFAULTOFFSET, pagesize = DEFAULTVOLUMEPAGESIZE, owned = false
     } = this.ctx.query;
-    if (this.ctx.user.id) {
+    if (this.ctx.session.user.id) {
       this.ctx.helper.createRes(403, 'permission denied ಠ益ಠ');
     }
     const response = await this.userService.getUserCollection(uid, offset, pagesize, owned);
@@ -51,71 +137,138 @@ class UserController extends Controller {
 
   async getMyCollection() {
     // 获取session里的uid
-    const uid = this.ctx.user.id;
+    const uid = this.ctx.session.user.id;
     if (uid) {
+      const {
+        offset, pagesize, owned 
+      } = this.ctx.query;
+      const response = await this.userService.getUserCollection(uid, offset, pagesize, owned);
+      this.ctx.helper.successRes('sucess',response);
+    } else{
       this.ctx.helper.createRes(404, 'User not find 凸(⊙▂⊙✖ )');
     }
-    const {
-      offset = DEFAULTOFFSET, pagesize = DEFAULTVOLUMEPAGESIZE, owned = false,
-    } = this.ctx.query;
-    const response = await this.userService.getUserCollection(uid, offset, pagesize, owned);
-    this.ctx.helper.successRes('sucess',response);
+
   }
+
+
   async getMyLike() {
-    const uid = this.ctx.user.id;
+    const uid = this.ctx.session.user.id;
     if (uid) {
-      this.ctx.helper.createRes(404, 'User not find 凸(⊙▂⊙✖ )');
       const {
         offset = DEFAULTOFFSET, pagesize = DEFAULTVOLUMEPAGESIZE
       } = this.ctx.query;
-      const response = await this.userService.getMyLike(uid, offset, pagesize, owned);
+      const response = await this.userService.getMyLike(uid, offset, pagesize);
       this.ctx.helper.successRes('sucess',response);
+    } else{
+      this.ctx.helper.createRes(404, 'User not find 凸(⊙▂⊙✖ )');
     }
   }
-  async starList() {
-    const uid = this.ctx.user.id;
+  async getMyStar() {
+    const uid = this.ctx.session.user.id;
     if (uid) {
-      this.ctx.helper.createRes(404, 'User not find 凸(⊙▂⊙✖ )');
+      
       const {
         offset = DEFAULTOFFSET, pagesize = DEFAULTVOLUMEPAGESIZE
       } = this.ctx.query;
-      const response = await this.userService.starList(uid, offset, pagesize, owned);
+      const response = await this.userService.getMyStar(uid, offset, pagesize);
       this.ctx.helper.successRes('sucess',response);
+    }else{
+      this.ctx.helper.createRes(404, 'User not find 凸(⊙▂⊙✖ )');
     }
   }
-  async UploadList() {
-    const uid = this.ctx.user.id;
+  async getMyUpload() {
+    const uid = this.ctx.session.user.id;
     if (uid) {
-      this.ctx.helper.createRes(404, 'User not find 凸(⊙▂⊙✖ )');
+
       const {
         offset = DEFAULTOFFSET, pagesize = DEFAULTVOLUMEPAGESIZE
       } = this.ctx.query;
-      const response = await this.userService.UploadList(uid, offset, pagesize, owned);
+      const response = await this.userService.getMyUpload(uid, offset, pagesize);
       this.ctx.helper.successRes('sucess',response);
+    } else{
+      this.ctx.helper.createRes(404, 'User not find 凸(⊙▂⊙✖ )');
+    }
+  }
+  async getMyCreate() {
+    const uid = this.ctx.session.user.id;
+    if (uid) {
+ 
+      const {
+        offset = DEFAULTOFFSET, pagesize = DEFAULTVOLUMEPAGESIZE
+      } = this.ctx.query;
+      const response = await this.userService.getMyCreate(uid, offset, pagesize);
+      this.ctx.helper.successRes('sucess',response);
+    }else{
+      this.ctx.helper.createRes(404, 'User not find 凸(⊙▂⊙✖ )');
     }
   }
   
-  // async checkPhone() {
-  //   const {phone}= this.ctx.params;
-  //   const response = await this.userService.checkPhone(phone);
-  //   if (response) {
-  //     this.ctx.helper.createRes(400,'此电话已经被使用')
-  //   } else {
-  //     this.ctx.helper.successRes('此电话未被使用')     
-  //   }
-  // }
+  
+
   
   async changephone() {
+    const {
+      phone,
+      password
+    } = this.ctx.request.body;
+    const uid = this.ctx.session.user.id;
+    uid || this.ctx.helper.createRes(203, 'User not login QAQ');
+    if(await this.userService.checkPhone(phone)){
+      this.ctx.helper.createRes(400, '此电话已被使用 凸(⊙▂⊙✖ )');
+    }else {
+      const response = await this.userService.changephone(phone, password, uid);
+      // console.log( this.response)
+      if(response == null){
+        this.ctx.helper.createRes(202, '用户名或密码错误(ﾟДﾟ;)')
+      }else{
+        this.ctx.session.user = response
+        this.ctx.helper.successRes('更改手机登录',response)
+      }
+    }
 
-  }
-
-  async remakespassword() {
-    
   }
 
   async changepassword() {
+    const {
+      newpassword,
+      password,
+    } = this.ctx.request.body;
+
+    const uid = this.ctx.session.user.id;
+    uid || this.ctx.helper.createRes(203, 'User not login QAQ');
+    
+      const response = await this.userService.changepassword(newpassword, password, uid);
+      // console.log( this.response)
+      if(response == null){
+        this.ctx.helper.createRes(202, '密码错误(ﾟДﾟ;)')
+      }else{
+        this.ctx.session.user = response
+        this.ctx.helper.successRes('更改密码登录',response)
+      }
     
   }
+
+  async remakespassword() {
+    const {
+      // username,
+      phone,
+      password,
+      // captcha,
+      sms
+    } = this.ctx.request.body;
+
+ if(await this.ctx.helper.checkSMS(sms)){
+      this.ctx.helper.createRes(400, '短信验证码错误');
+    }else  if( !await this.userService.checkPhone(phone)){
+      this.ctx.helper.createRes(400, '此电话未注册 凸(⊙▂⊙✖ )');
+    }else{
+      const response = await this.userService.remakespassword( phone, password);
+      this.ctx.helper.successRes('sucess',response); 
+    }
+    
+  }
+
+
 
   async login() {
     const {
@@ -164,7 +317,7 @@ class UserController extends Controller {
   }
 
   async logout() {
-    this.ctx.logout();
+    this.ctx.session.user = {}
     this.ctx.helper.successRes('sucess',this.ctx.session);
   }
 
@@ -172,12 +325,17 @@ class UserController extends Controller {
     const {
       vid
     } = this.ctx.request.body;
-    const uid = this.ctx.user.id;
-    const response = await this.userService.addCollectionVolume(uid, vid);
-    if (!response[response.length - 1]) { // check is new record
-      this.ctx.helper.createRes(412, ' volume has been collected Orz');
-    } else {
-      this.ctx.helper.createRes(200, 'collection success QwQ');
+    const uid = this.ctx.session.user.id;
+    if(uid){
+
+      const response = await this.userService.addCollectionVolume(uid, vid);
+      if (!response[response.length - 1]) { // check is new record
+        this.ctx.helper.createRes(412, ' volume has been collected Orz');
+      } else {
+        this.ctx.helper.createRes(200, 'collection success QwQ');
+      }
+    } else{
+      this.ctx.helper.createRes(404, 'User not find 凸(⊙▂⊙✖ )');
     }
   }
 
@@ -185,16 +343,21 @@ class UserController extends Controller {
     const {
       vid
     } = this.ctx.params;
-    const uid = this.ctx.user.id;
-    const response = await this.userService.deleteCollectionVolume(uid, vid);
-    if (response) {
-      this.ctx.helper.createRes(200, 'Delete success QwQ');
+    const uid = this.ctx.session.user.id;
+    if(uid){
 
-    } else {
-      this.ctx.helper.createRes(409, 'Delete err Orz  ');
-
+      const response = await this.userService.deleteCollectionVolume(uid, vid);
+      if (response) {
+        this.ctx.helper.createRes(200, 'Delete success QwQ');
+  
+      } else {
+        this.ctx.helper.createRes(409, 'Delete err Orz  ');
+  
+      }
+      this.ctx.helper.successRes('sucess',response);
+    } else{
+      this.ctx.helper.createRes(404, 'User not find 凸(⊙▂⊙✖ )');
     }
-    this.ctx.helper.successRes('sucess',response);
   }
   
   async addStarUser() {
